1. clean a lot of c++ warnings, mark functions const where appropriate. 2. clean js a little to ensure distinct objects are created for procedureMgr, tableMgr 3. introduce new way to bind types to c++ driver - initially only support varBinary via tm.sqlTypes.asVarBinary(null) to ensure a null is bound to varbinary type. 4. add 2 unit tests for above.

+++ src/Utility.cpp
@@ -55,7 +55,7 @@ namespace mssql
         {
             // length includes null terminator
             messageBuffer.resize(length);
-            ::WideCharToMultiByte(CP_UTF8, 0, input, -1, messageBuffer.data(), messageBuffer.size(), nullptr, nullptr);
+            ::WideCharToMultiByte(CP_UTF8, 0, input, -1, messageBuffer.data(), static_cast<int>(messageBuffer.size()), nullptr, nullptr);
         }
         return string(messageBuffer.data());
     }
